Updated files to match player points (changed to int)

+++ src/main/scala/model/Player.scala
@@ -1,3 +1,3 @@
 package model
 
-case class Player(id: Int, name: String, email: String, passwordHash: String, points: Double)
+case class Player(id: Int, name: String, email: String, passwordHash: String, points: Int)
